backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
dlib==19.24.2
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.0.3
Pillow==10.0.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.24
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.7
wrapt==2.0.1
